Multiple changes, added timers, added memcpy

+++ common/const.h
@@ -8,6 +8,6 @@
 #define MIN_CLIENT_MSG_LEN 13
 #define MAX_CLIENT_MSG_LEN 33
 
-using Coord = std::pair<uint32_t, uint32_t>;
+using Coord = std::pair<double, double>;
 
 #endif //SCREEN_WORMS_CONST_H
